Restructure project to a better folder structure
