+++ src/main/java/client/TCPClient.java
@@ -1,4 +1,4 @@
-package greenhouse.entities;
+package client;
 
 import javax.crypto.Cipher;
 import javax.crypto.SecretKey;
