--- src/main/java/greenhouse/entities/ServerSetupException.java
@@ -1,7 +1,0 @@
-package greenhouse.entities;
-
-public class ServerSetupException extends RuntimeException {
-    public ServerSetupException(String message) {
-        super(message);
-    }
-}
