--- src/main/java/greenhouse/entities/Toggleable.java
@@ -1,8 +1,0 @@
-package greenhouse.entities;
-
-public interface Toggleable {
-
-  void start();
-
-  void stop();
-}
