+++ src/main/java/greenhouse/entities/sensors/PHSensor.java
@@ -1,7 +1,7 @@
 package greenhouse.entities.sensors;
 
-import greenhouse.entities.Soil;
 import greenhouse.entities.SoilSubscriber;
+import greenhouse.logic.Soil;
 
 /**
  * Represents a sensor reading the ph of soil.
