--- src/main/java/greenhouse/entities/sensors/SensorNotYetActiveException.java
@@ -1,7 +1,0 @@
-package greenhouse.entities.sensors;
-
-public class SensorNotYetActiveException extends RuntimeException {
-    public SensorNotYetActiveException(String message) {
-        super(message);
-    }
-}
