+++ src/test/java/greenhouse/entities/SoilTest.java
@@ -3,6 +3,8 @@ package greenhouse.entities;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
+import greenhouse.logic.Soil;
+
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
